Log email errors
