+++ lib/teiserver_web/controllers/account/session_controller.ex
@@ -199,7 +199,7 @@ defmodule TeiserverWeb.Account.SessionController do
         |> render("forgot_password.html")
 
       true ->
-        case Teiserver.Account.Emails.send_password_reset(user) do
+        case Teiserver.EmailHelper.send_password_reset(user) do
           :ok ->
             conn
             |> put_flash(:success, "Password reset email sent out")
